refactor the old model so that using parts of it will be easier
