+++ mypy.ini
@@ -1,6 +1,10 @@
 [mypy]
 namespace_packages = True
+disallow_subclassing_any = False
 plugins=mypy_zope:plugin
+
+# making our way to 'strict'
+warn_return_any = True
 
 strict_optional = True
 warn_no_return = True
